Backend_Main_SpringBoot. Corregidas y modificadas tablas de migración relacionadas con Suscripciones para cambiar id por bigint. Añadida configuración LoggerFactory en application.properties

+++ backend_main_springboot/src/main/resources/db/migration/V6__create_relationship_feature_tables.sql
@@ -11,7 +11,7 @@ CREATE TABLE profile_logros (
 CREATE TABLE pagos (
     id int PRIMARY KEY AUTO_INCREMENT,
     profile_id bigint,
-    suscripcion_id int,
+    suscripcion_id bigint,
     monto decimal(10,2),
     fecha_pago timestamp DEFAULT CURRENT_TIMESTAMP,
     fecha_inicio date,
